some import statement fixing and initial editcontrols refactoring
