+++ src/utils/SelectOptions.js
@@ -1,8 +1,8 @@
 /** @module SingleEdit/SelectOptions */
-import * as Contents from './EditContents.js';
-import * as Grouping from './Grouping.js';
-import * as Notification from '../utils/Notification.js';
-import { SplitHandler } from './StaffTools.js';
+import * as Contents from '../SquareEdit/EditContents.js';
+import * as Grouping from '../SquareEdit/Grouping.js';
+import * as Notification from './Notification.js';
+import { SplitHandler } from '../SquareEdit/StaffTools.js';
 const $ = require('jquery');
 
 /**
