+++ src/utils/SelectTools.js
@@ -2,8 +2,8 @@
 
 import * as Color from '../utils/Color.js';
 import { updateHighlight } from '../DisplayPanel/DisplayControls.js';
-import * as Grouping from './Grouping.js';
-import { Resize } from './StaffTools.js';
+import * as Grouping from '../SquareEdit/Grouping.js';
+import { Resize } from '../SquareEdit/StaffTools.js';
 import * as SelectOptions from './SelectOptions.js';
 
 const d3 = require('d3');
